Chore - rename files - add file handler - rename root service folder
